nic_router: fix Eth dest MAC in domain broadcasts
When the router receives a packet for a domain-local IP that is not the routers
IP at that domain it broadcasts this packet to all interfaces of the domain.
Unfortunately, it does only adapt the Ethernet source MAC in this case. The
Ethernet destination MAC stays the same, the same as the new source MAC, the
router MAC.

It stands to reason whether the router should instead start an ARP request to
find the correct MAC (and thereby the correct interface which would prevent
broadcasting) or set the destination MAC to the Ethernet broadcast MAC.

This is the fastest version by simply using a broadcast MAC.

Ref #3625