sculpt: make graph features configurable
This patch make the component graph better reusable for the phone
version of Sculpt. In the phone version, the '+' menu does not exist. So
we need to omit the corresponding button. Furthermore, the storage
dialog is presented in a dedicate section of the GUI instead of
presenting it inside the 'storage' graph node.