part_block: mediate access to partitions
This commit changes how the partition multiplexer operates. Previously
it created on Block connection to the driver while providing a session
for each partition to its clients. The block connection as well as the
sessions used there own, differently sized bulk-buffer for transferring
data. While the data coming from a driver was normally transfered via
DMA, this data was again copied into the client bulk-buffer.

In contrast to that it now solely mediates access to the block-device
and the partitions on top of that. It parses the partition-table and
forwards an augmented Block session request, which contains the offset
where the partition starts as well as the number of blocks covered by
the partition, to the block-device driver. The client collaborates
with the driver directly now. The 'io_buffer' attribute is made
obsolete by this change as the client now specifies the I/O buffer
size itself.

The Block session used to parse the partition-table is ephemeral in
nature and is closed after the parsing was completed.

Issue #5479.