hw: group-forward and warp-limiting in scheduler
Group fast-worward:
When a group was unready for a relatively short time,
e.g. waiting for cross-core IPC, it should not be penalized
by fast-forwarding its vtime to min_vtime and letting the group
wait for every group with a larger warp value. To cirumvent
this, we assume the group had been scheduled the entire time
instead of being unready. If the notional vtime is still smaller
than min_vtime, we let the group continue with this vtime.

Warp limiting:
Adds a maximum warp execution time.

Moreover, this commit adds an explanatory comment to the
scheduler class declaration to facilitate getting started with
the scheduler algorithm.

Issue #5117